refactor(feed): split feed page into UA and CI versions
