+++ src/pages/feed/index.ts
@@ -1,0 +1,3 @@
+import { Feed } from './_shared';
+
+export const GET = await Feed();
